데미지 및 이동속도 장비 스탯 연동 완료
체력만 하면 끝

+++ PangeaOdessey/Assets/Scripts/Item.cs
@@ -16,6 +16,6 @@ public class Item : ScriptableObject
     public EquipmentType equipmentType;
 
     public int Hp;
-    public int Damage;
-    public int Speed;
+    public float Damage;
+    public float Speed;
 }
